+++ addons/mx/cfdi/food_vouchers.go
@@ -6,7 +6,7 @@ import (
 	"github.com/invopop/gobl/cal"
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/num"
-	"github.com/invopop/gobl/regimes/mx/sat"
+	"github.com/invopop/gobl/regimes/mx"
 	"github.com/invopop/validation"
 )
 
@@ -103,7 +103,7 @@ func (fve *FoodVouchersEmployee) Validate() error {
 	return validation.ValidateStruct(fve,
 		validation.Field(&fve.TaxCode,
 			validation.Required,
-			validation.By(sat.ValidateTaxCode),
+			validation.By(mx.ValidateTaxCode),
 		),
 		validation.Field(&fve.CURP,
 			validation.Required,