+++ addons/mx/cfdi/tax_combo.go
@@ -1,18 +1,18 @@
 package cfdi
 
 import (
-	"github.com/invopop/gobl/regimes/mx/sat"
+	"github.com/invopop/gobl/regimes/mx"
 	"github.com/invopop/gobl/tax"
 )
 
 func normalizeTaxCombo(tc *tax.Combo) {
 	var k tax.ExtValue
 	switch tc.Category {
-	case sat.TaxCategoryISR:
+	case mx.TaxCategoryISR:
 		k = "001"
-	case tax.CategoryVAT, sat.TaxCategoryRVAT:
+	case tax.CategoryVAT, mx.TaxCategoryRVAT:
 		k = "002"
-	case sat.TaxCategoryIEPS, sat.TaxCategoryRIEPS:
+	case mx.TaxCategoryIEPS, mx.TaxCategoryRIEPS:
 		k = "003"
 	default:
 		return