+++ regimes/mx/corrections.go
@@ -1,15 +1,10 @@
-package sat
+package mx
 
 import (
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/tax"
 )
-
-// CorrectionDefinitions provides the array of correction definitions that apply to SAT
-func CorrectionDefinitions() []*tax.CorrectionDefinition {
-	return correctionDefinitions
-}
 
 var correctionDefinitions = []*tax.CorrectionDefinition{
 	{
@@ -18,7 +13,7 @@ var correctionDefinitions = []*tax.CorrectionDefinition{
 			bill.InvoiceTypeCreditNote,
 		},
 		Stamps: []cbc.Key{
-			StampUUID,
+			StampSATUUID,
 		},
 	},
 }