+++ regimes/pl/invoices.go
@@ -83,7 +83,7 @@ func (v *invoiceValidator) commercialCustomer(value interface{}) error {
 }
 
 func (v *invoiceValidator) preceding(value interface{}) error {
-	obj, ok := value.(*bill.Preceding)
+	obj, ok := value.(*org.DocumentRef)
 	if !ok || obj == nil {
 		return nil
 	}