+++ CHANGELOG.md
@@ -30,6 +30,9 @@ Finally, the `draft` flag has been removed from the header, and much more emphas
 - `pt`: moved SAF-T specific extensions to `addons/pt/saft`.
 - `it`: moved SDI and FatturaPA extensions to `addons/it/sdi` with key `it-sdi-v1`.
 - `gr`: moved MyDATA to `addons/gr/mydata`, key `gr-mydata-v1`.
+- `bill.Preceding`: replaced with `org.DocumentRef`.
+- `bill.Invoice`: Ordering now using arrays of `org.DocumentRef`.
+- `bill.Invoice`: `series` and `code` now use `cbc.Code` and normalization instead of the independent invoice code.
 
 ### Added
 
@@ -39,6 +42,7 @@ Finally, the `draft` flag has been removed from the header, and much more emphas
 - `tax`: `Scenario` now has `Filter` property to set a code function.
 - `tax`: `AddonDef` provides support for defining addon extension packs.
 - `gr`: `gr-mydata-invoice-type` extension with related tags and scenarios.
+- `org`: `DocumentRef` consolidates references to previous documents in a single place.
 
 ## [v0.115.1]